+++ R/ggrisktable.R
@@ -10,6 +10,7 @@ NULL
 #'  \bold{percentage} of subjects at risk by time, respectively. Use "abs_pct"
 #'  to show both absolute number and percentage.
 #'@param title the title of the plot.
+#'@param xlog logical value. If TRUE, x axis is tansformed into log scale.
 #'@param y.text logical. Default is TRUE. If FALSE, the table y axis. tick
 #'  labels will be hidden.
 #'@param y.text.col logical. Default value is FALSE. If TRUE, the table tick
@@ -35,8 +36,9 @@ NULL
 #'
 #'@export
 ggrisktable <- function (fit, data = NULL, type = c("absolute", "percentage", "abs_pct", "nrisk_cumcensor", "nrisk_cumevents"),
-                         color = "black", palette = NULL, break.time.by = NULL,  xlim = NULL,
+                         color = "black", palette = NULL, break.time.by = NULL,  xlim = NULL, xscale = 1,
                          title = NULL, xlab = "Time", ylab = "Strata",
+                         xlog = FALSE,
                          legend = "top",
                          legend.title = "Strata", legend.labs = NULL, y.text = TRUE, y.text.col = TRUE, fontsize = 4.5,
                          ggtheme = theme_light(), ...)
@@ -44,7 +46,6 @@ ggrisktable <- function (fit, data = NULL, type = c("absolute", "percentage", "a
 
   if(!inherits(fit, "survfit"))
     stop("Can't handle an object of class ", class(fit))
-  if(is.null(xlim)) xlim <- c(0, max(fit$time))
   .check_legend_labs(fit, legend.labs)
 
   type <- match.arg(type)
@@ -62,8 +63,10 @@ ggrisktable <- function (fit, data = NULL, type = c("absolute", "percentage", "a
   data <- .get_data(fit, data = data)
 
   # Define time axis breaks
-  if(is.null(break.time.by)) times <- .get_default_breaks(fit$time)
-  else times <- seq(0, max(c(fit$time, xlim)), by = break.time.by)
+  xmin <- ifelse(xlog, min(c(1, fit$time)), 0)
+  if(is.null(xlim)) xlim <- c(xmin, max(fit$time))
+  times <- .get_default_breaks(fit$time, .log = xlog)
+  if(!is.null(break.time.by) &!xlog) times <- seq(0, max(c(fit$time, xlim)), by = break.time.by)
 
   survsummary <- .get_timepoints_survsummary(fit, data, times)
 
@@ -95,10 +98,15 @@ ggrisktable <- function (fit, data = NULL, type = c("absolute", "percentage", "a
     ggtheme +
     scale_y_discrete(breaks = as.character(levels(survsummary$strata)), labels = yticklabs ) +
     coord_cartesian(xlim = xlim) +
-    scale_x_continuous(breaks = times) +
     labs(title = title, x = xlab, y = ylab, color = legend.title, shape = legend.title)
 
+  p <- .set_risktable_gpar(p, ...)
   p <- ggpubr::ggpar(p, legend = legend, palette = palette,...)
+
+  xticklabels <- .format_xticklabels(labels = times, xscale = xscale)
+  if(!xlog) p <- p + ggplot2::scale_x_continuous(breaks = times, labels = xticklabels)
+  else p <- p + ggplot2::scale_x_continuous(breaks = times, trans = "log10", labels = xticklabels)
+
 
   if(!y.text) p <- .set_large_dash_as_ytext(p)
 
@@ -113,6 +121,20 @@ ggrisktable <- function (fit, data = NULL, type = c("absolute", "percentage", "a
 
   p
 
+}
+
+# For backward compatibility
+# Specific graphical params to risk.table
+.set_risktable_gpar <- function(p,  ...){
+  extra.params <- list(...)
+  ggpubr:::.labs(p,
+        font.main = extra.params$font.risk.table.title,
+        font.x = extra.params$font.risk.table.x,
+        font.y = extra.params$font.risk.table.y,
+        submain = extra.params$risk.table.subtitle,
+        caption = extra.params$risk.table.caption,
+        font.submain = extra.params$font.risk.table.subtitle,
+        font.caption = extra.params$font.risk.table.caption)
 }