+++ R/pairwise_survdiff.R
@@ -45,7 +45,8 @@ pairwise_survdiff <- function(formula, data, p.adjust.method = "BH", na.action, 
 {
   if(missing(na.action)) na.action <- options()$na.omit
   group_var <- attr(stats::terms(formula), "term.labels")
-  group <- data[, group_var]
+  group <- unlist(data[, group_var])
+  if(!is.factor(group)) group <- as.factor(group)
   ngroup <- length(levels(group))
 
   DNAME <- paste(deparse(substitute(data)), "and", group_var)