+++ R/utilities.R
@@ -97,8 +97,9 @@ GeomConfint <- ggplot2::ggproto('GeomConfint', ggplot2::GeomRibbon,
 # Compute default axis breaks as ggplot2
 #-------------------------------------
 # Return a vector of axis labels
-.get_default_breaks <- function(x){
-  scales::extended_breaks()(x)
+.get_default_breaks <- function(x, .log = FALSE){
+  if(!.log) scales::extended_breaks()(x)
+  else scales::log_breaks()(x)
 }
 
 
@@ -209,5 +210,34 @@ GeomConfint <- ggplot2::ggproto('GeomConfint', ggplot2::GeomRibbon,
 .set_large_dash_as_ytext <- function(ggp){
   ggp + theme(axis.text.y = element_text(size = 50, vjust = 0.35),
         axis.ticks.y = element_blank())
+}
+
+
+# Transform x-axis labels according to the scale see ggsurvplot().
+#
+# labels: numeric vector (x-axis labels)
+#xscale: numeric or character values (see
+# gsurvplot). If numeric, the value is used to divide the labels on the x axis.
+# For example, a value of 365.25 will give labels in years instead of the
+# original days. If character, allowed options include one of c("d_m", "d_y",
+# "m_d", "m_y", "y_d", "y_m"), where d = days, m = months and y = years. For
+# example, xscale = "d_m" will transform labels from days to months; xscale =
+# "m_y", will transform labels from months to years.
+.format_xticklabels <- function(labels, xscale){
+
+  # 1 year = 365.25 days
+  # 1 month = 365.25/12 = 30.4375 days
+  if(is.numeric(xscale)) xtrans <- 1/xscale
+  else
+    xtrans <- switch(xscale,
+                     d_m = 12/365.25,
+                     d_y = 1/365.25,
+                     m_d = 365.25/12,
+                     m_y = 1/12,
+                     y_d = 365.25,
+                     y_m = 12,
+                     1
+    )
+  round(labels*xtrans,2)
 }